requirements/dev.txt
Django
django-river
six
django-codemirror2
ldap3
river-admin
